pyproject.toml
[project]
name = "gold-miner"
version = "0.1.0"
description = "Lightweight agent for MaxCompute SQL analysis with memory and skills"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "pandas>=2.0.0",
  "pyodps>=0.12.0",
  "tabulate>=0.9.0",
  "tavily-python>=0.5.0",
  "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
]

[project.scripts]
gold-miner = "gold_miner.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
